+++ CHANGELOG.md
@@ -1,5 +1,17 @@
 # Changelog
 
+## 06/05/2023 - v3.13.2
+
+- **UVJ:**
+  - (Add) Properties: BottomWaitTimeBeforeCure, WaitTimeBeforeCure, BottomWaitTimeAfterCure, WaitTimeAfterCure, BottomWaitTimeAfterLift, WaitTimeAfterLift
+  - (Add) Vendor key to the configuration to be able to save custom key-values from other softwares (#687)
+- (Add) Layers properties: `CompletionTime`, `CompletionTimeStr`, `StartTime`, `StartTimeStr`, `EndTime`, `EndTimeStr` (#698)
+- (Add) PrusaSlicer printer: Creality Halot Mage and Mage Pro
+- (Add) Tool - Blur: Stack blur
+- (Change) Tool - Timelapse: Increase wait time from 100s to 1000s maximum
+- (Change) Settings: Allow to combine start maximized with restore windows position and size (#695)
+- (Fix) File formats: Parse transition layer count from layers now only take into account decreasing times related to the previous layer
+
 ## 27/04/2023 - v3.13.1
 
 - (Change) `Layer.IsBottomLayer` no longer calculate the value using the position of the layer, a new property `IsBottomLayerByHeight` is now used to get that result